+++ riak/transports/tcp/codec.py
